renamed fake_io to io, use std::io if std is enabled
