+++ algebra/src/biginteger/mod.rs
@@ -1,6 +1,6 @@
 use crate::{
     bytes::{FromBytes, ToBytes},
-    fake_io::{Read, Result as IoResult, Write},
+    io::{Read, Result as IoResult, Write},
     fields::BitIterator,
     UniformRand, Vec,
 };
