+++ algebra/src/curves/mnt6/g1.rs
@@ -7,7 +7,7 @@ use crate::{
         short_weierstrass_projective::{GroupAffine, GroupProjective},
         AffineCurve, PairingCurve, PairingEngine,
     },
-    fake_io::{Result as IoResult, Write},
+    io::{Result as IoResult, Write},
     field_new,
     fields::mnt6::{Fq, Fq3, Fq6, Fr},
 };
