+++ algebra/src/curves/models/bls12/g1.rs
@@ -5,7 +5,7 @@ use crate::{
         short_weierstrass_jacobian::{GroupAffine, GroupProjective},
         AffineCurve,
     },
-    fake_io::{Result as IoResult, Write},
+    io::{Result as IoResult, Write},
 };
 use num_traits::Zero;
