+++ algebra/src/curves/models/short_weierstrass_projective.rs
@@ -1,6 +1,6 @@
 use crate::{
     curves::models::SWModelParameters as Parameters,
-    fake_io::{Read, Result as IoResult, Write},
+    io::{Read, Result as IoResult, Write},
     CanonicalDeserialize, CanonicalSerialize, UniformRand, Vec,
 };
 use core::{
