+++ algebra/src/fields/models/fp6_3over2.rs
@@ -1,5 +1,5 @@
 use crate::{
-    fake_io::{Read, Result as IoResult, Write},
+    io::{Read, Result as IoResult, Write},
     CanonicalDeserialize, CanonicalSerialize, SerializationError, UniformRand,
 };
 use core::{
